pyproject.toml
[tool.poetry]
name = "bee-the-indolent"
version = "0.1.0"
description = "Silly discord bot with fun commands for use in lounge servers"
authors = ["pixDeVl <93949551+pixDeVl@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.3.2"
pillow = "^10.1.0"
logger = "^1.4"
asyncio = "^3.4.3"
python-dotenv = "^1.0.0"
regex = "^2023.10.3"
datetime = "^5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
